adicionada a funcao de carrinho de compras e simulacao de pagamento utilizando cartao de credito

+++ Projeto2_AED1/GerenciadorDeCadastroDeUsuario.cs
@@ -136,7 +136,7 @@ namespace Projeto2_AED1
 
         private static bool ChecaSeStringEhCompostaPorNumeroInteiro(string str)
         {
-            return int.TryParse(str, out _);
+            return Int64.TryParse(str, out _);
         }
     }
 }
